[#2] Restructure graphics
We now return a plotly figure object instead of just plotting it.

Improved test for graphics.

Removed some files and renamed functions.
